# precalculate SHA256 hashes. Speed-up: 1.2x