--- a/Cargo.toml
+++ b/Cargo.toml
@@ -37,6 +37,7 @@ num-integer = "0.1.45"
 chia-bls = "0.4.0"
 sha2 = "0.10.8"
 openssl = { version = "=0.10.55", features = ["vendored"], optional = true }
+hex-literal = "=0.4.1"
 # for secp sigs
 k256 = { version = "0.13.1", features = ["ecdsa"] }
 p256 = { version = "0.13.2", features = ["ecdsa"] }