@arvidn 2024-02-22
TODO:

* [ ] extend unit test to cover all optimized cases + some
* [ ] apply the same optimization to the other tree-hash function as well