@richardkiss 2024-02-22
Be sure to do a test for hashing the two values `01 00`, ie. the one-byte but non-minimal 0 integer value. We want to make sure it's different from `01`, ie. the minimal 0 integer value.