@richardkiss 2024-02-22
Remarkable speed-up!